fix: update formatDuration function to support long form output and adjust DurationInput component accordingly
(cherry picked from commit 35326f1d3f0ac1b58171464e587fdabba9cc01d7)
